+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### PaymentSheet
 * [Added] The ability to customize the appearance of the PaymentSheet using `PaymentSheet.Appearance`.
 * [Added] Support for collecting payments from customers in 54 additional countries within PaymentSheet. Most of these countries are located in Africa and the Middle East.
+* [Added] `affirm` and `AUBECSDebit` payment methods are now available in PaymentSheet
 
 ## 22.3.2 2022-05-18
 ### CardScan